requirements.txt
streamlit
streamlit-autorefresh
fastapi
requests
python-dotenv
pytest
uvicorn
mcp[cli]
git+https://github.com/modelcontextprotocol/python-sdk.git
yfinance
altair
feedparser